+++ axelor-production/src/main/java/com/axelor/apps/production/service/operationorder/OperationOrderServiceImpl.java
@@ -20,6 +20,7 @@ package com.axelor.apps.production.service.operationorder;
 import com.axelor.apps.base.db.DayPlanning;
 import com.axelor.apps.base.service.BarcodeGeneratorService;
 import com.axelor.apps.base.service.weeklyplanning.WeeklyPlanningService;
+import com.axelor.apps.production.db.Machine;
 import com.axelor.apps.production.db.ManufOrder;
 import com.axelor.apps.production.db.OperationOrder;
 import com.axelor.apps.production.db.ProdHumanResource;
@@ -82,7 +83,7 @@ public class OperationOrderServiceImpl implements OperationOrderService {
             manufOrder,
             prodProcessLine.getPriority(),
             prodProcessLine.getWorkCenter(),
-            prodProcessLine.getWorkCenter(),
+            prodProcessLine.getWorkCenter().getMachine(),
             prodProcessLine);
 
     return Beans.get(OperationOrderRepository.class).save(operationOrder);
@@ -93,7 +94,7 @@ public class OperationOrderServiceImpl implements OperationOrderService {
       ManufOrder manufOrder,
       int priority,
       WorkCenter workCenter,
-      WorkCenter machineWorkCenter,
+      Machine machineWorkCenter,
       ProdProcessLine prodProcessLine)
       throws AxelorException {
 
@@ -113,7 +114,7 @@ public class OperationOrderServiceImpl implements OperationOrderService {
             OperationOrderRepository.STATUS_DRAFT,
             prodProcessLine);
 
-    this._createHumanResourceList(operationOrder, machineWorkCenter);
+    this._createHumanResourceList(operationOrder, workCenter);
 
     return Beans.get(OperationOrderRepository.class).save(operationOrder);
   }